requirements.txt
django
gunicorn
django-heroku
psycopg2
djangorestframework
djangorestframework-simplejwt
git+git://github.com/johnfraney/wkhtmltopdf-pack.git
pdfkit==0.6.1
bottle